+++ src/routes.php
@@ -1,8 +1,9 @@
 <?php
 
+use Butler\Service\Http\Controllers\ConsumersController;
+use Butler\Service\Http\Controllers\FrontController;
 use Butler\Service\Http\Controllers\GraphqlController;
 use Butler\Service\Http\Controllers\HealthController;
-use Butler\Service\Http\Controllers\FrontController;
 use Illuminate\Support\Facades\Route;
 
 Route::middleware('web')->group(function () {
@@ -22,4 +23,9 @@ Route::middleware(['api', 'auth'])->group(function () {
         config('butler.service.routes.graphql', '/graphql'),
         GraphqlController::class
     )->name('graphql');
+
+    Route::get(
+        config('butler.service.routes.consumers', '/consumers'),
+        ConsumersController::class
+    )->middleware('can:view-consumers')->name('consumers');
 });