+++ tests/config/butler.php
@@ -37,6 +37,7 @@ return [
             'front' => '/',
             'graphql' => '/graphql',
             'health' => '/health',
+            'consumers' => '/consumers',
         ],
 
         'health' => [